Move JsonSchemaService to domain module (#113)
- Moves the `JsonSchemaService` to the domain module
- Schema validations are only executed on domain level entities and inside domain level repositories
- Rename `/domain/errors` to `/domain/schema` – the entities that were under "errors" were related with schema validation. Since JsonSchemaService is now under `/domain/schema` the "schema" name seems more fitting
