+++ src/domain/backbone/backbone.repository.spec.ts
@@ -1,7 +1,7 @@
 import { faker } from '@faker-js/faker';
 import { HttpException } from '@nestjs/common';
-import { JsonSchemaService } from '../../common/schemas/json-schema.service';
-import { ValidationErrorFactory } from '../errors/validation-error-factory';
+import { JsonSchemaService } from '../schema/json-schema.service';
+import { ValidationErrorFactory } from '../schema/validation-error-factory';
 import backboneFactory from '../balances/entities/__tests__/backbone.factory';
 import { ITransactionApi } from '../interfaces/transaction-api.interface';
 import { ITransactionApiManager } from '../interfaces/transaction-api.manager.interface';
