+++ src/domain/backbone/backbone.repository.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { DefinedError, ValidateFunction } from 'ajv';
-import { JsonSchemaService } from '../../common/schemas/json-schema.service';
-import { ValidationErrorFactory } from '../errors/validation-error-factory';
+import { JsonSchemaService } from '../schema/json-schema.service';
+import { ValidationErrorFactory } from '../schema/validation-error-factory';
 import { backboneSchema } from '../balances/entities/schemas/backbone.schema';
 import { ITransactionApiManager } from '../interfaces/transaction-api.manager.interface';
 import { Backbone } from './entities/backbone.entity';
