+++ src/domain/schema/validation-error-factory.ts
