+++ src/routes/cache-hooks/cache-hooks.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { CacheHooksController } from './cache-hooks.controller';
-import { JsonSchemaService } from '../../common/schemas/json-schema.service';
+import { JsonSchemaService } from '../../domain/schema/json-schema.service';
 import { CacheHooksService } from './cache-hooks.service';
 
 @Module({
